Temp

+++ package_app.sh
@@ -1,14 +1,16 @@
 #!/bin/sh
 
 set -x
+set -e
 
 if [ -z "$1" ] || [ -z "$2" ]; then
 	echo "Usage: sh package_app.sh path/to/buck platform-flavors"
 	exit 1
 fi
 
-PLATFORM_FLAVORS="$2"
 BUCK="$1"
+BUCK_OPTIONS="$2"
+PLATFORM_FLAVORS="$3"
 
 BUILD_PATH="build"
 APP_NAME="Telegram"
@@ -26,19 +28,19 @@ mkdir -p "$DSYMS_DIR"
 
 cp "buck-out/gen/App/AppPackage#$PLATFORM_FLAVORS.ipa" "$IPA_PATH.original"
 rm -rf "$IPA_PATH.original.unpacked"
-rm "$BUILD_PATH/${APP_NAME}_signed.ipa"
+rm -f "$BUILD_PATH/${APP_NAME}_signed.ipa"
 mkdir -p "$IPA_PATH.original.unpacked"
 unzip "$IPA_PATH.original" -d "$IPA_PATH.original.unpacked/"
 rm "$IPA_PATH.original"
 
 UNPACKED_PATH="$IPA_PATH.original.unpacked"
-APP_PATH="$UNPACKED_PATH/Payload/App.app"
+APP_PATH="$UNPACKED_PATH/Payload/Telegram.app"
 FRAMEWORKS_DIR="$APP_PATH/Frameworks"
 
 rm -rf "$IPA_PATH.original.unpacked/SwiftSupport/iphoneos/"*
 rm -rf "$FRAMEWORKS_DIR"/*
 
-for DEPENDENCY in $($BUCK query "kind('apple_library|apple_binary', deps('//App:App#$PLATFORM_FLAVORS', 1))"); do
+for DEPENDENCY in $(${BUCK} query "kind('apple_library|apple_binary', deps('//App:Telegram#$PLATFORM_FLAVORS', 1))" ${BUCK_OPTIONS}); do
 	case "$DEPENDENCY" in 
 		*"#"*)
 			;;
@@ -49,8 +51,10 @@ for DEPENDENCY in $($BUCK query "kind('apple_library|apple_binary', deps('//App:
 	DEPENDENCY_PATH=$(echo "$DEPENDENCY" | sed -e "s#^//##" | sed -e "s#:#/#")
 	DEPENDENCY_NAME=$(echo "$DEPENDENCY" | sed -e "s/#.*//" | sed -e "s/^.*\://")
 	DYLIB_PATH="buck-out/gen/$DEPENDENCY_PATH/lib$DEPENDENCY_NAME.dylib"
-	TARGET_DYLIB_PATH="$FRAMEWORKS_DIR/lib$DEPENDENCY_NAME.dylib"
-	cp "$DYLIB_PATH" "$TARGET_DYLIB_PATH"
+	if [ -f "$DYLIB_PATH" ]; then
+		TARGET_DYLIB_PATH="$FRAMEWORKS_DIR/lib$DEPENDENCY_NAME.dylib"
+		cp "$DYLIB_PATH" "$TARGET_DYLIB_PATH"
+	fi
 	DSYM_PATH="buck-out/gen/$(echo "$DEPENDENCY" | sed -e "s/#/#apple-dsym,/" | sed -e "s#^//##" | sed -e "s#:#/#").dSYM"
 	cp -f -r "$DSYM_PATH" "$DSYMS_DIR/"
 done
@@ -59,7 +63,7 @@ for LIB in $(ls "$FRAMEWORKS_DIR"/*.dylib); do
 	strip -S -T "$LIB"
 done
 
-xcrun swift-stdlib-tool --scan-folder "$IPA_PATH.original.unpacked/Payload/App.app" --scan-folder "$IPA_PATH.original.unpacked/Payload/App.app/Frameworks" --strip-bitcode --platform iphoneos --copy --destination "$IPA_PATH.original.unpacked/SwiftSupport/iphoneos"
+xcrun swift-stdlib-tool --scan-folder "$IPA_PATH.original.unpacked/Payload/Telegram.app" --scan-folder "$IPA_PATH.original.unpacked/Payload/Telegram.app/Frameworks" --strip-bitcode --platform iphoneos --copy --destination "$IPA_PATH.original.unpacked/SwiftSupport/iphoneos"
 
 for LIB in $(ls "$IPA_PATH.original.unpacked/SwiftSupport/iphoneos/"*.dylib); do
 	codesign --remove-signature "$LIB"
